+++ src/main/java/com/example/echo/EchoApplication.java
@@ -1,0 +1,15 @@
+package com.example.echo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@SpringBootApplication
+@EnableJpaAuditing
+public class EchoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(EchoApplication.class, args);
+	}
+
+}
