+++ src/main/java/com/example/echo/domain/member/entity/Role.java
@@ -1,0 +1,6 @@
+package com.example.echo.domain.member.entity;
+
+public enum Role {
+    ADMIN,
+    USER
+}
