+++ src/main/java/com/example/echo/domain/petition/entity/Petition.java
@@ -1,5 +1,6 @@
-package com.example.nbe2_2_team03.entity;
+package com.example.echo.domain.petition.entity;
 
+import com.example.echo.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.*;
