--- src/main/java/com/example/nbe2_2_team03/Nbe22Team03Application.java
@@ -1,13 +1,0 @@
-package com.example.nbe2_2_team03;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Nbe22Team03Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Nbe22Team03Application.class, args);
-	}
-
-}
