--- src/main/java/com/example/nbe2_2_team03/entity/Role.java
@@ -1,6 +1,0 @@
-package com.example.nbe2_2_team03.entity;
-
-public enum Role {
-    ADMIN,
-    USER
-}
