+++ src/test/java/com/example/echo/EchoApplicationTests.java
@@ -1,10 +1,10 @@
-package com.example.nbe2_2_team03;
+package com.example.echo;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class Nbe22Team03ApplicationTests {
+class EchoApplicationTests {
 
 	@Test
 	void contextLoads() {
